feat(channel): add per-channel rate limiting
- configurable requests/window via channel other-settings, enforced on Redis (sliding window) and in-memory backends
- selection skips rate-limited channels and auto-groups fall through to the next healthy group
- pinned/specific channels, async image executors and task retries surface 429 (channel:rate_limited)
- channel form UI + 7-locale i18n; Go and frontend tests

+++ middleware/model-rate-limit.go
@@ -104,7 +104,7 @@ func redisRateLimitHandler(duration int64, totalMaxCount, successMaxCount int) g
 				ctx,
 				totalKey,
 				limiter.WithCapacity(int64(totalMaxCount)*duration),
-				limiter.WithRate(int64(totalMaxCount)),
+				limiter.WithRate(float64(totalMaxCount)),
 				limiter.WithRequested(duration),
 			)
